+++ services/sponsor/internal/service/service.go
@@ -47,12 +47,12 @@ func (s *Service) Sponsor(ctx context.Context, request *models.Sponsorship, owne
 		return nil, 0, err
 	}
 
-	amount := float32(res.Tier * 20)
+	amount := int64(res.Tier * 20)
 
 	receipt_id, err := s.makePayment(ctx, &models.PaymentRequest{
 		UserID:         owner,
 		UserEmail:      email,
-		IdempotencyKey: GenerateIdempotencyKey(owner, amount, int32(res.ID), "S"),
+		IdempotencyKey: GenerateIdempotencyKey(owner, (float32)(amount), int32(res.ID), "S"),
 		Amount:         amount,
 		PaymentDescription: fmt.Sprintf("SPONSOR %d FOR RESTAURANT %d WITH TIER %d UNTIL %s",
 			res.ID, request.ID, res.Tier, FormatTime(res.Until)),
